runner/requirements.txt
python-dateutil==2.8.1
falcon==2.0.0
falcon-cors==1.1.7
fleep==1.0.1
gunicorn==20.0.4
jsonschema==3.2.0
requests>=2.23.0
youtube-dl==2020.12.14
falcon_multipart>=0.2.0
psutil>=5.7.0
tweepy>=3.9.0
numpy>=1.19.0
pillow>=10.0.0
rq==1.16.2
redis==5.0.0
werkzeug==2.3.7
ipfs-api==0.2.3
dateparser==1.2.0
ipfsApi==0.4.4
yt_dlp==2025.3.21
